add sorting + refactorings
