+++ src/components/property-tree/property-tree.tsx
@@ -26,7 +26,8 @@ export const PropertyTree = ({ asset, childAssets }: PropertyTreeProps): JSX.Ele
   const [childNodes, setChildNodes] = useState<TreeAsset[]>([]);
 
   useEffect(() => {
-    setChildNodes(addChildrenAssets(childAssets, asset.id));
+    const childrenAssets = addChildrenAssets(childAssets, asset.id)
+    setChildNodes(childrenAssets);
   }, [childAssets, asset]);
 
   useEffect(() => {
