+++ src/components/related-assets/types.ts
@@ -1,0 +1,5 @@
+export interface RelatedAsset {
+  id: string;
+  name: string;
+  type: string;
+}
